+++ R/venndir-conversion.R
@@ -29,6 +29,9 @@
 #' setlist <- counts2setlist(combo)
 #' setlist;
 #' 
+#' # turn off warnings about GEOS
+#' options("warn"=-1)
+#' 
 #' # Venn diagram
 #' venndir(setlist)
 #'